Enhance YouTube scene functionality and improve user experience
- Added a cancel button to the YouTube scene steps to allow users to abort operations.
- Updated message handling to skip processing if the user is already in a scene.
- Improved formatting and consistency in constant definitions and message responses.
- Introduced a new step for audio downloading in the YouTube service.
- Enhanced error handling and user feedback during video and audio download processes.

+++ src/scenes/youtube/steps/youtube_step_3.ts
@@ -7,24 +7,37 @@ import { config } from "@/config/config";
 import { Input } from "telegraf";
 import { startContinuousAction } from "@/utils/continuousAction";
 import { IMyContext } from "@/interfaces/IMyContext";
+import { CANCEL_BUTTON_CALLBACK } from "@/scenes/youtube/constants";
 
 export const youtubeStep3 = async (ctx: TYoutubeSceneContext) => {
-    await ctx.answerCbQuery();
+  await ctx.answerCbQuery();
 
-    if (ctx.has(callbackQuery("data"))) {
-        const height = ctx.callbackQuery.data;
-        await ctx.deleteMessage();
-        const message = await ctx.sendMessage(MESSAGES.INFO.DOWNLOADING_VIDEO);
-        const folderName = await ctx.scene.session.service.downloadVideo(height);
-        startContinuousAction(ctx as IMyContext, "upload_video");
-        // get first file in folder
-        const filePath = path.join(config.downloadDir, folderName, fs.readdirSync(path.join(config.downloadDir, folderName))[0]);
-        await ctx.deleteMessage(message.message_id);
-        const videoStream = fs.createReadStream(filePath);
-        await ctx.replyWithVideo(Input.fromReadableStream(videoStream));
-        // delete the folder inside config.downloadDir
-        fs.rmSync(path.join(config.downloadDir, folderName), { recursive: true });
+  if (ctx.has(callbackQuery("data"))) {
+    const data = ctx.callbackQuery.data;
+
+    // Handle cancel button
+    if (data === CANCEL_BUTTON_CALLBACK) {
+      await ctx.editMessageText(MESSAGES.INFO.SCENE_CANCELLED);
+      return ctx.scene.leave();
     }
 
-    return ctx.scene.leave();
-}
+    const height = data;
+    await ctx.deleteMessage();
+    const message = await ctx.sendMessage(MESSAGES.INFO.DOWNLOADING_VIDEO);
+    const folderName = await ctx.scene.session.service.downloadVideo(height);
+    startContinuousAction(ctx as IMyContext, "upload_video");
+    // get first file in folder
+    const filePath = path.join(
+      config.downloadDir,
+      folderName,
+      fs.readdirSync(path.join(config.downloadDir, folderName))[0]
+    );
+    await ctx.deleteMessage(message.message_id);
+    const videoStream = fs.createReadStream(filePath);
+    await ctx.replyWithVideo(Input.fromReadableStream(videoStream));
+    // delete the folder inside config.downloadDir
+    fs.rmSync(path.join(config.downloadDir, folderName), { recursive: true });
+  }
+
+  return ctx.scene.leave();
+};
